conver_confirm_modal is migrated
